+++ SimG4CMS/Calo/src/HGCSD.cc
@@ -32,7 +32,7 @@ HGCSD::HGCSD(G4String name, const DDCompactView & cpv,
 	     const SensitiveDetectorCatalog & clg, 
 	     edm::ParameterSet const & p, const SimTrackManager* manager) : 
   CaloSD(name, cpv, clg, p, manager,
-         p.getParameter<edm::ParameterSet>("HGCSD").getParameter<int>("TimeSliceUnit"),
+         p.getParameter<edm::ParameterSet>("HGCSD").getParameter<double>("TimeSliceUnit"),
          p.getParameter<edm::ParameterSet>("HGCSD").getParameter<bool>("IgnoreTrackID")), 
   numberingScheme(0) {
